+++ internal/db/database.go
@@ -141,7 +141,6 @@ func validateConfig(cfg *config.DatabaseConfig) error {
 func autoMigrate(db *gorm.DB) error {
 	return db.AutoMigrate(
 		&user.User{},
-		&user.WechatAccount{},
 		&user.UserChannel{},
 		&user.LLMConfig{},
 		&conversation.Message{},